Move remaining Contract, ContractHelper code to shared partition.
