+++ src/System.Private.CoreLib/shared/System/Runtime/CompilerServices/ContractHelper.cs
